fragment_gym_requirements.txt
setuptools==57.5.0
wheel==0.37.1
shapely==2.0.1
absl-py==1.4.0
appdirs==1.4.4
attrdict3==2.0.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
docker-pycreds==0.4.0
Farama-Notifications==0.0.4
filelock==3.12.0
fonttools==4.39.4
gdown==5.1.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.19.1
google-auth-oauthlib==1.0.0
gprof2dot==2022.7.29
grpcio==1.54.2
gym==0.21.0
gymnasium==0.28.1
idna==3.4
importlib-metadata==4.13.0
jax-jumpy==1.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
lit==16.0.5
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.2
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.23.2
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sb3-contrib==1.8.0
sentry-sdk==1.26.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stable-baselines3==1.8.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
torch==2.0.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.4
Werkzeug==2.3.4
zipp==3.15.0
huggingface-hub==0.23.4
wasabi==1.1.3